Move Crafty.scene into a new separate file
Crafty.scene should not be in drawing.js.  This commit moves it into a new file, "scene.js".
